agno-api/frontend/requirements.txt
streamlit
requests
sseclient-py